blurrer/requirements.txt
pillow==12.1.0
requests==2.32.5
Flask==3.1.2